requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
autoevals==0.0.129
braintrust_core==0.0.59
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
comm==0.2.3
cryptography==45.0.5
cyclopts==3.22.2
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastmcp==2.10.6
filelock==3.18.0
fire==0.7.0
frozenlist==1.7.0
fsspec==2025.3.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.67.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
ibm-cos-sdk==2.14.2
ibm-cos-sdk-core==2.14.2
ibm-cos-sdk-s3transfer==2.14.2
ibm_watson_machine_learning==1.0.368
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.74
langchain-mcp-adapters==0.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.13
litellm==1.74.9.post1
llama_api_client==0.1.2
llama_stack==0.2.18
llama_stack_client==0.2.18
lomond==0.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp>=1.12.4
mcp-streamablehttp-proxy==0.2.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
openapi-pydantic==0.5.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
packaging==25.0
pandas==2.1.4
parso==0.8.4
peft==0.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
polyleven==0.9.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary>=2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.7.0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymilvus>=2.4.10
pypdf==6.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.29
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
trl==0.21.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
